refactor: update import paths, enhance logging, and switch to ChatOpenAI model
