+++ Pipfile
@@ -9,14 +9,13 @@ pytest = "*"
 pypdf2 = "*"
 langgraph = "*"
 langchain = "*"
-huggingface-hub = "*"
 ipython = "*"
-langchain-huggingface = "*"
 google-cloud-bigquery = "*"
 pytest-cov = "*"
 python-dotenv = "*"
 streamlit = "*"
 langchain-community = "*"
+
 
 [dev-packages]
 pandas = "*"
