+++ config.yaml
@@ -1,3 +1,5 @@
+log_level: INFO
+log_format: "%(asctime)s - %(name)s - %(levelname)s - %(message)s"
 bigquery:
   project_id: astrafy
   dataset_id: smart_scholar
@@ -20,5 +22,5 @@ bigquery:
       mode: "REPEATED"
       description: "List of keywords."
 inputs:
-  llm_model: 'TEST'
-  temperature: 0.7
+  llm_model: 'gpt-3.5-turbo' #gpt-3.5-turbo or TEST
+  temperature: 0.001
