+++ src/constants.py
@@ -6,7 +6,8 @@ PROJECT_ID='12345'
 DATASET_ID='abcde'
 TABLE_NAME='AcademicPapers'
 
+LLM_API_TOKEN = 'OPENAI_API_KEY'
+STORAGE_PROVIDER_CREDENTIALS = 'GOOGLE_APPLICATION_CREDENTIALS'
 
-DEFAULT_API_KEY='None'
 DEFAULT_TEMPERATURE = 0.5
-DEFAULT_LLM_MODEL = "gpt2"
+DEFAULT_LLM_MODEL = 'gpt-3.5-turbo'
